@@ -9,24 +9,23 @@
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@iconify/json": "^2.2.260",
-    "@sveltejs/adapter-auto": "^3.2.5",
-    "@sveltejs/adapter-node": "^5.2.7",
+    "@iconify/json": "^2.2.264",
+    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/adapter-node": "^5.2.9",
     "@sveltejs/adapter-vercel": "^4.0.5",
-    "@sveltejs/kit": "^2.7.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/kit": "^2.7.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/canvas-confetti": "^1.6.4",
     "@types/howler": "^2.2.12",
     "@types/marked": "^4.3.2",
     "@types/three": "^0.143.2",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.4.47",
-    "svelte": "^4.2.19",
-    "svelte-preprocess": "^5.1.4",
-    "svelte-windicss-preprocess": "^4.2.8",
-    "tailwindcss": "^3.4.13",
+    "svelte": "^5.1.3",
+    "svelte-preprocess": "^6.0.3",
+    "tailwindcss": "^3.4.14",
     "typescript": "^5.6.3",
-    "vite": "^5.4.9"
+    "vite": "^5.4.10"
   },
   "type": "module",
   "dependencies": {
@@ -37,6 +36,6 @@
     "marked": "^4.3.0",
     "mei": "github:ninjawarrior1337/mei",
     "three": "^0.143.0",
-    "unplugin-icons": "^0.14.15"
+    "unplugin-icons": "^0.19.3"
   }
 }
