@@ -1,2 +1,7 @@
-<img class="lg:order-last mx-12 transform-gpu py-16 aspect-square w-[350px] lg:w-[450px] scale-[175%] translate-y-12" src="/logo/logo2024.svg" alt="big logo"/>
+<img
+  class="lg:order-last mx-12 transform-gpu py-16 aspect-square w-[350px] lg:w-[450px] scale-[175%] translate-y-12"
+  src="/logo/logo2024.svg"
+  alt="big logo"
+/>
 <!-- <canvas class="lg:order-last" bind:this={canvas} width="450" height="450" /> -->
+<script lang="ts"></script>
